Test Protocol client
This test client is used to validate the communication channel and protocol between this client and language server. The client can send request and notification to server based on the language protocol and expect response for the request. Server can also send specific notification to client. Any custom protocol in addition to the language server protocol can be built on to od existing channel.

Mocha can used to run tests for this protocol client.