Change TextDocument's getText to take an optional range
TextDocument's getText() function has been changed to take an
optional range in the form of getText(range?: Range). This allows
clients to extract a substring within the document instead of the
entire document. Ranges that start at a position smaller than the
document will be adjusted to the beginning of the document. Likewise,
ranges that end beyond the length of the document will be adjusted to
end within the bounds of the document.

Signed-off-by: Remy Suen <remy.suen@gmail.com>