Add middleware to intercept diagnostics from the server
Clients may want to inspect diagnostics received from the server
before letting VS Code handle and display them. It is not possible to
use onNotification because the existing handler will be replaced.
Thus, we need to introduce a new middleware for the language client
to route the server's textDocument/publishDiagnostics notification
to. This will allow the client can do what it needs to do before then
forwarding it off to VS Code.

Signed-off-by: Remy Suen <remy.suen@gmail.com>