Fix documentLink/resolve by introducing a data field
The current definition of a DocumentLink makes resolving a link with
documentLink/resolve essentially impossible. Having only a range
field filled out is not enough information for a language server to
deduce what content or file the link is about. To correct this, a new
data field has been introduced to DocumentLink that servers can fill
in with some data on the initial fulfillment of a
textDocument/documentLink request so that subsequent
documentLink/resolve requests will have some contextual information
available to resolve the link.

Signed-off-by: Remy Suen <remy.suen@gmail.com>