Replace symlinking in Travis build with npm scripts
In npm 5 (which is included in Node.js 8), symbolic links in the
node_modules folder are now replaced by what's in the npm registry.
As a result, symbolic links between the language server packages now
need to be created after npm install has completed.

Pushing the linking down to being a npm script instead of performing
this as part of the CI build also allows it to be reused for local
development. This has the advantage of making local development more
closely match what is happening in the CI builds.

Signed-off-by: Remy Suen <remy.suen@gmail.com>