Add support for preparing a rename request
Implemented a new textDocument/prepareRename request. This request
allows a language server to resolve and validate the current cursor
location before a rename is actually attempted. This allows clients to
interrupt a rename request if for example the user's text cursor is
currently within a language keyword and as such cannot be renamed.

Signed-off-by: Remy Suen <remy.suen@gmail.com>