Remove per request cancellation file at the end of the request, if possible.
This is a best-effort approach, so it doesn't guarantee all files will be deleted at the end.

For example, if the client-side tries to raise cancellation right after the server responded, but before the client-side responded to the reply, then it could create a cancellation file that no one is looking for.

Another case would be if cancellation is raised, but no one in the server-side checked the cancellation, then the server wouldn't know that cancellation is raised, and the cancellation file created won't be deleted.

But the owner of the JSON-RPC connection owns the cancellation file folder, so he should take care of resources left at the end of the session.