fixed two issues found
First, I forgot to put clean up at the end of a test.
Second, I found an issue where the event dispose method throws an exception if it is called after Emitter is disposed.

This can now happen for CancellationTokenSource since messageConnection holds onto disposable from event subscription to call clean up after request is completed (succeeded or not).

But most of the time, that is after cancellation is called for the source, which disposes the Emitter.  so we need to handle this case.