Add final newlines
This enables the VSCode option `files.insertFinalNewline` and reformats
all files. In most cases this means the final newline is added. In some
cases trailing whitespace was stripped as well.

The main reason to prefer inserting a final newline, is because POSIX
defines a text file as a file made of lines, where each line ends with a
newline character. This includes the final line.

Most tools write final newlines, i.e. npm, tsc. Git and related tools
tend to render a missing newline as a sort of warning. This often leads
to annoying diffs if the newline is missing.