protocol: Fix typo in WorkspaceCapabilities.diagnostic field name
This commit fixes a typo in the name of `diagnostic` field on client's workspace capabilities. Put another way, it removes the trailing `s` from the name of `diagnostic` field to align it with the LSP spec.

This turned out to be an issue with tsgo, which is generating it's JSON deserialization based on metaModel.json. This then led to it being incompatible with Zed, which is a non-VSC client.

Related:
- https://github.com/microsoft/typescript-go/issues/1396
- https://github.com/zed-extensions/tsgo/issues/6