Drop textDocument/didClose if dropping textDocument/didOpen (#1682)
* Drop textDocument/didClose if dropping textDocument/didOpen

When using delayOpenNotifications=true, the open notifications are queued and dropped if a close notification immediately follows. However, the original implementation still sent the close notification. This does not match the spec:

> A close notification requires a previous open notification to be sent

This change also drops the close notification if the open notification was dropped.

Fixes https://github.com/microsoft/vscode-languageserver-node/issues/1681

* Fix comment