Do not recreate logging channel after stop (#1694)
We only clean up the logging channel once after
the client transitions to the Stopped state. We
should make sure it is not created again afterwards
as this means the output channel will leak. Let's
add a helper method shouldLogToOutputChannel()
which will return False when we're in the Stopped
state and the channel has been cleaned up already
and use it in the logging methods.

Fixes #1693

Co-authored-by: Dirk Bäumer <dirkb@microsoft.com>